requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
pdf2image>=1.17.0
pandas>=2.0.0
werkzeug>=3.0.0
requests>=2.31.0
numpy>=1.24.0
PyPDF2>=3.0.0
openai>=1.75.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
Pillow>=10.0.0